Fixed problem causing book objects to duplicate
When the user enters a book entry for the second time, it would double
the book, continuing each time a book submitted.

Removing onclick action from the button in the html and moving listener
from the function and instead being it's own arrow function and calls
to the function fixed the problem.